Create separate Scrubber library and scrub error.log
