+++ lib/ts/Diags.cc
@@ -175,8 +175,8 @@ Diags::~Diags()
     stderr_log = nullptr;
   }
 
-  for (auto s : scrubs) {
-    delete s;
+  if (scrubber) {
+    delete scrubber;
   }
 
   ats_free((void *)base_debug_tags);
@@ -884,86 +884,6 @@ Diags::set_stderr_output(const char *stderr_path)
   return ret;
 }
 
-void
-Diags::scrub_add(const char *pattern, const char *replacement)
-{
-  pcre *re;
-  const char *error;
-  int erroroffset;
-  Scrub *s;
-
-  // compile the regular expression
-  re = pcre_compile(pattern, 0, &error, &erroroffset, NULL);
-  if (!re) {
-    log_log_error("[Warning]: Unable to compile PCRE scrubbing pattern\n");
-    log_log_error("[Warning]: Scrubbing pattern failed at offset %d: %s.\n", erroroffset, error);
-    return;
-  }
-
-  // add the scrub pattern to our list
-  s              = new Scrub;
-  s->pattern     = ats_strdup(pattern);
-  s->replacement = ats_strdup(replacement);
-  s->compiled_re = re;
-  scrubs.push_back(s);
-}
-
-char *
-Diags::scrub_buffer(const char *buffer, Scrub *scrub) const
-{
-  int num_matched;
-  char *scrubbed;
-  char *scrubbed_idx;
-
-  // execute regex
-  num_matched = pcre_exec(scrub->compiled_re, nullptr, buffer, strlen(buffer), 0, 0, scrub->ovector, scrub->OVECCOUNT);
-  if (likely(num_matched < 0)) {
-    switch (num_matched) {
-    case PCRE_ERROR_NOMATCH:
-      return ats_strdup(buffer);
-    default:
-      log_log_error("[Warning] PCRE matching error %d\n", num_matched);
-      break;
-    }
-  }
-
-  // guaranteed to be big enough
-  scrubbed     = static_cast<char *>(ats_malloc(strlen(buffer) + strlen(scrub->replacement) + 1));
-  scrubbed_idx = scrubbed;
-
-  // copy over all the stuff before the captured substing
-  memcpy(scrubbed_idx, buffer, scrub->ovector[0]);
-  scrubbed_idx += scrub->ovector[0];
-
-  // copy over the scrubbed stuff
-  int replacement_len = strlen(scrub->replacement);
-  memcpy(scrubbed_idx, scrub->replacement, replacement_len);
-  scrubbed_idx += replacement_len;
-
-  // copy over everything after the scrubbed stuff
-  int trailing_len = strlen(buffer + scrub->ovector[1]);
-  memcpy(scrubbed_idx, buffer + scrub->ovector[1], trailing_len);
-  scrubbed_idx += trailing_len;
-
-  // nul terminate
-  *scrubbed_idx = '\0';
-
-  return scrubbed;
-}
-
-char *
-Diags::scrub_buffer(const char *buffer, const std::vector<Scrub *> &scrubs) const
-{
-  // apply every Scrub in the vector, in order
-  char *scrubbed = ats_strdup(buffer);
-  for (auto s : scrubs) {
-    char *new_scrubbed = scrub_buffer(scrubbed, s);
-    ats_free(scrubbed);
-    scrubbed = new_scrubbed;
-  }
-  return scrubbed;
-}
-
 /*
  * Helper function that rebinds stdout to specified file descriptor
  *
@@ -1011,7 +931,7 @@ Diags::vprintline(FILE *fp, char *buffer, va_list ap) const
       int r     = vsnprintf(buf, buf_size, buffer, ap);
       // if it's fully written, then we can break
       if (r < buf_size) { // we break on encoding errors (ie r < 0)
-        char *scrubbed = scrub_buffer(buf, scrubs);
+        char *scrubbed = scrubber->scrub_buffer(buf);
         fprintf(fp, scrubbed);
         ats_free(scrubbed);
         ats_free(buf);
