+++ lib/ts/Diags.h
@@ -34,16 +34,15 @@
 #ifndef __DIAGS_H___
 #define __DIAGS_H___
 
-#include <vector>
 #include <stdarg.h>
 #include "ink_mutex.h"
-#include "pcre.h"
 #include "Regex.h"
 #include "ink_apidefs.h"
 #include "ContFlags.h"
 #include "ink_inet.h"
 #include "BaseLogFile.h"
 #include "SourceLocation.h"
+#include "Scrubber.h"
 
 #define DIAGS_MAGIC 0x12345678
 #define BYTES_IN_MB 1000000
@@ -92,26 +91,6 @@ struct DiagsConfigState {
   // this is static to eliminate many loads from the critical path
   static bool enabled[2];                    // one debug, one action
   DiagsModeOutput outputs[DiagsLevel_Count]; // where each level prints
-};
-
-struct Scrub {
-  ~Scrub()
-  {
-    if (pattern) {
-      ats_free(const_cast<char *>(pattern));
-    }
-    if (replacement) {
-      ats_free(const_cast<char *>(replacement));
-    }
-    if (compiled_re) {
-      pcre_free(const_cast<pcre *>(compiled_re));
-    }
-  }
-  static const int OVECCOUNT = 30;
-  const char *pattern;
-  const char *replacement;
-  const pcre *compiled_re;
-  int ovector[OVECCOUNT];
 };
 
 //////////////////////////////////////////////////////////////////////////////
@@ -248,28 +227,11 @@ public:
    * This really shouldn't be called by anyone other than DiagsConfig b/c there's inherent race conditions.
    */
   void
-  scrub_config(bool enable)
+  scrub_config(bool enable, const char *config)
   {
     scrub_enabled = enable;
+    scrubber      = new Scrubber(config);
   }
-  /*
-   * Add another expression to scrub from diagnostic logs.
-   *
-   * This really shouldn't be called by anyone other than DiagsConfig b/c there's inherent race conditions.
-   */
-  void scrub_add(const char *pattern, const char *replacement);
-
-  /*
-   * Heap allocates an identical buffer that is scrubbed.
-   * Caller should ats_free.
-   */
-  char *scrub_buffer(const char *buffer, Scrub *scrub) const;
-
-  /*
-   * Heap allocates an identical buffer that is scrubbed with multiple Scrubs.
-   * Caller should ats_free.
-   */
-  char *scrub_buffer(const char *buffer, const std::vector<Scrub *> &scrubs) const;
 
   const char *base_debug_tags;  // internal copy of default debug tags
   const char *base_action_tags; // internal copy of default action tags
@@ -281,7 +243,7 @@ private:
 
   // Scrubbing variables -- replaces predefined regexp matches with the replacement string
   bool scrub_enabled = false;
-  std::vector<Scrub *> scrubs;
+  Scrubber *scrubber = nullptr;
 
   // These are the default logfile permissions
   int diags_logfile_perm  = -1;
