+++ proxy/logging/LogObject.cc
@@ -124,6 +124,13 @@ LogObject::LogObject(const LogFormat *format, const char *log_dir, const char *b
   ink_assert(b);
   SET_FREELIST_POINTER_VERSION(m_log_buffer, b, 0);
 
+  // set up scrubbing object
+  char *scrubs = REC_ConfigReadString("proxy.config.diags.scrubs");
+  if (scrubs) {
+    scrub_enabled = true;
+    scrubber      = new Scrubber(scrubs);
+  }
+
   _setup_rolling(rolling_enabled, rolling_interval_sec, rolling_offset_hr, rolling_size_mb);
 
   Debug("log-config", "exiting LogObject constructor, filename=%s this=%p", m_filename, this);
@@ -159,6 +166,9 @@ LogObject::LogObject(LogObject &rhs)
     add_loghost(host);
   }
 
+  scrub_enabled = rhs.scrub_enabled;
+  scrubber      = new Scrubber(*(rhs.scrubber));
+
   // copy gets a fresh log buffer
   //
   LogBuffer *b = new LogBuffer(this, Log::config->log_buffer_size);
@@ -182,6 +192,11 @@ LogObject::~LogObject()
       m_host_list.clear();
     }
   }
+
+  if (scrubber) {
+    delete scrubber;
+  }
+
   ats_free(m_basename);
   ats_free(m_filename);
   ats_free(m_alt_filename);
@@ -578,6 +593,9 @@ LogObject::log(LogAccess *lad, const char *text_entry)
   } else if (lad) {
     bytes_needed = m_format->m_field_list.marshal_len(lad);
   } else if (text_entry) {
+    if (scrub_enabled) {
+      text_entry = scrubber->scrub_buffer(text_entry);
+    }
     bytes_needed = LogAccess::strlen(text_entry);
   }
