+++ proxy/logging/LogObject.h
@@ -299,6 +299,10 @@ private:
   unsigned m_buffer_manager_idx;
   LogBufferManager *m_buffer_manager;
 
+  // scrubbing data -- helps us scrub buffers of sensitive data
+  bool scrub_enabled = false;
+  Scrubber *scrubber;
+
   void generate_filenames(const char *log_dir, const char *basename, LogFileFormat file_format);
   void _setup_rolling(Log::RollingEnabledValues rolling_enabled, int rolling_interval_sec, int rolling_offset_hr,
                       int rolling_size_mb);
