+++ proxy/shared/DiagsConfig.h
@@ -37,7 +37,6 @@ struct DiagsConfig {
   ~DiagsConfig();
 
 private:
-  void config_scrubbing(Diags *d, const char *config);
   bool callbacks_established;
   BaseLogFile *diags_log;
